Move actions from plan module to provider module
